--- a/src/main/java/com/timgroup/statsd/BufferPool.java
+++ b/src/main/java/com/timgroup/statsd/BufferPool.java
@@ -3,33 +3,33 @@ package com.timgroup.statsd;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 
-import java.util.Queue;
+import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ArrayBlockingQueue;
 
 public class BufferPool {
-    private final Queue<ByteBuffer> pool;
+    private final BlockingQueue<ByteBuffer> pool;
     private final int size;
 
 
-    BufferPool(final int poolSize, int bufferSize, final boolean direct) {
+    BufferPool(final int poolSize, int bufferSize, final boolean direct) throws InterruptedException {
 
         size = poolSize;
         pool = new ArrayBlockingQueue<ByteBuffer>(poolSize);
         for (int i=0; i<size ; i++) {
             if (direct) {
-                pool.offer(ByteBuffer.allocateDirect(bufferSize));
+                pool.put(ByteBuffer.allocateDirect(bufferSize));
             } else {
-                pool.offer(ByteBuffer.allocate(bufferSize));
+                pool.put(ByteBuffer.allocate(bufferSize));
             }
         }
     }
 
-    ByteBuffer borrow() {
-        return pool.poll();
+    ByteBuffer borrow() throws InterruptedException {
+        return pool.take();
     }
 
-    void put(ByteBuffer buffer) {
-        pool.offer(buffer);
+    void put(ByteBuffer buffer) throws InterruptedException {
+        pool.put(buffer);
     }
 
     int available() {