--- a/src/main/java/com/timgroup/statsd/BufferPool.java
+++ b/src/main/java/com/timgroup/statsd/BufferPool.java
@@ -3,8 +3,8 @@ package com.timgroup.statsd;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 
-import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ArrayBlockingQueue;
+import java.util.concurrent.BlockingQueue;
 
 public class BufferPool {
     private final BlockingQueue<ByteBuffer> pool;
@@ -15,7 +15,7 @@ public class BufferPool {
 
         size = poolSize;
         pool = new ArrayBlockingQueue<ByteBuffer>(poolSize);
-        for (int i=0; i<size ; i++) {
+        for (int i = 0; i < size ; i++) {
             if (direct) {
                 pool.put(ByteBuffer.allocateDirect(bufferSize));
             } else {