--- a/src/main/java/com/timgroup/statsd/BufferPool.java
+++ b/src/main/java/com/timgroup/statsd/BufferPool.java
@@ -1,6 +1,5 @@
 package com.timgroup.statsd;
 
-import java.io.IOException;
 import java.nio.ByteBuffer;
 
 import java.util.concurrent.ArrayBlockingQueue;