--- a/src/main/java/com/timgroup/statsd/InvalidMessageException.java
+++ b/src/main/java/com/timgroup/statsd/InvalidMessageException.java
@@ -1,7 +1,7 @@
 package com.timgroup.statsd;
 
 /**
- * Signals that we've been passed a message that's invalid and won't be sent
+ * Signals that we've been passed a message that's invalid and won't be sent.
  *
  * @author Taylor Schilling
  */
@@ -11,7 +11,7 @@ public class InvalidMessageException extends RuntimeException {
     private final String invalidMessage;
 
     /**
-     * Creates an InvalidMessageException with a specified detail message and the invalid message itself
+     * Creates an InvalidMessageException with a specified detail message and the invalid message itself.
      *
      * @param detailMessage a message that details why the invalid message is considered so
      * @param invalidMessage the message deemed invalid