--- a/src/main/java/com/timgroup/statsd/NonBlockingStatsDClientBuilder.java
+++ b/src/main/java/com/timgroup/statsd/NonBlockingStatsDClientBuilder.java
@@ -114,6 +114,7 @@ public class NonBlockingStatsDClientBuilder {
 
     /**
      * NonBlockingStatsDClient factory method.
+     * @return the built NonBlockingStatsDClient.
      */
     public NonBlockingStatsDClient build() throws StatsDClientException {
         if (addressLookup != null) {