--- a/src/main/java/com/timgroup/statsd/StatsDClientException.java
+++ b/src/main/java/com/timgroup/statsd/StatsDClientException.java
@@ -3,7 +3,7 @@ package com.timgroup.statsd;
 
 /**
  * Signals that an exception has occurred when trying to start the
- * StatsD client
+ * StatsD client.
  * 
  * @author Tom Denley
  *