The goal of this PR is to improve client performance by removing client blocking calls, replacing them with non-blocking variants and letting the caller deal with failure. 

### Benchmark

A full-blast performance test has been added to try to understand the performance gains yielded by the `ConcurrentLinkedList` vs `LinkedBlockingQueue` (current implementation). The test is very simple (see: 3dddaef), using 10 threads to submit metrics in a busy loop with the statsd client. On the other end a dummy statsd server.

The following numbers have been obtained on a late 2015 13-inch Macbook Pro (3.1GHz i7 and 16GB RAM).

**1024 element queue**:
- `ConcurrentLinkedList`
```
Running com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest
Dec 10, 2019 6:30:10 PM com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest perf_test
INFO: Messages at server: 12470160
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 30.159 sec
```
- `LinkedBlockingQueue`
```
Running com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest
Dec 10, 2019 6:26:36 PM com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest perf_test
INFO: Messages at server: 1670500
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 31.128 sec
```

**2048 element queue**:
- `ConcurrentLinkedList`
```
Running com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest
Dec 10, 2019 6:20:38 PM com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest perf_test
INFO: Messages at server: 11791287
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 30.195 sec
```
- `LinkedBlockingQueue`
```
Running com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest
Dec 10, 2019 6:23:07 PM com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest perf_test
INFO: Messages at server: 2399716
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 31.095 sec
```

**4096 element queue**:
- `ConcurrentLinkedList`
```
Running com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest
Dec 10, 2019 6:38:35 PM com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest perf_test
INFO: Messages at server: 11745778
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 30.178 sec
```
- `LinkedBlockingQueue`
```
Running com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest
Dec 10, 2019 6:39:58 PM com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest perf_test
INFO: Messages at server: 2606678
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 31.09 sec
```

**Integer.MAX_VALUE queue**
- `ConcurrentLinkedList`
```
Running com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest
Dec 10, 2019 6:46:40 PM com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest perf_test
INFO: Messages at server: 1566180
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 51.564 sec
```
- `LinkedBlockingQueue`
```
Running com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest
Dec 10, 2019 6:44:54 PM com.timgroup.statsd.NonBlockingStatsDClientMaxPerfTest perf_test
INFO: Messages at server: 7216704
Tests run: 1, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 58.014 sec
```

#### Conclusions

While there are some very clear gains using the `ConcurrentLinkedList`, particularly when comparing the best performance achieved, the diminishing returns with larger queue sizes for the `ConcurrentLinkedList` scenarios is remarkable and hard to explain. Particularly given that the changes made to the message building and actual channel(socket) writing logic are minimal.

While the `LinkedBlockingQueue` scenario steadily provides increased performance with larger queues, we have the reverse behavior for our `ConcurrentLinkedList`. Note that due to implementation details there are a few operations in the `ConcurrentLinkedList` which are not, as one would expect, constant time O(1) (for instance, `size()`). However, we tried to steer clear from these O(n) methods, so it's currently challenging to explain the difference in performance and the issues we encountered. 

(to be continued....)